requirements.txt
uvloop ~= 0.17.0; sys_platform != 'win32'
hikari[speedups] ~= 2.0.0.dev112
hikari-lightbulb ~= 2.2.5
lavasnek-rs ~= 0.1.0a5
aiohttp ~= 3.8.3
python-dotenv[cli] ~= 0.21.0
yt-dlp ~= 2022.11.11
beautifulsoup4 ~= 4.11.1
